engine: extract esm/{constellations,sky_text,tours/text_object}.js
